Renamed text() to dynamicText(), added simpleText()
The original text glyph function is now dynamicText(), and there is a
simpleText() function that doesn't have the cute text detail changing
functionality.
