+++ src/glyphs/text.ts
@@ -1,94 +1,14 @@
 import { Annotation } from "../annotations/annotation";
 import { Chart } from "../charts/chart";
-import * as d3 from "d3";
 import { GlyphConfig } from "../glyph-utilities/glyph-config";
-import { generateId } from "../utilities/id-generation";
-import { bind } from "../glyph-utilities/bind";
 import {
   GlyphModifier,
   GlyphModifierConfig,
   GlyphProperty,
 } from "../glyph-utilities/glyph-modifier";
 
-const textMap: Map<string, string[]> = new Map();
-const thresholdMap: Map<string, number[]> = new Map();
-
 /**
- * A helper function that decides what text to render in a text glyph given the Chart's zoom level.
  * @internal
- * @param a
- * @param c
- */
-export function selectText(a: Annotation, c: Chart<any>): string {
-  let thresholds = thresholdMap.get(a.id);
-  if (thresholds === undefined) {
-    console.error(
-      `text thresholds undefined for annotation: ${a.id}, returning empty string`
-    );
-    return "";
-  }
-
-  let text = textMap.get(a.id);
-  if (text === undefined) {
-    console.error(
-      `text undefined for annotation: ${a.id} returning empty string`
-    );
-    return "";
-  }
-
-  let i = 0;
-  let viewWidth = c.getSemanticViewRange().width;
-  for (const thresh of thresholds) {
-    if (viewWidth <= thresh) {
-      return text[i];
-    }
-    i++;
-  }
-  return "";
-}
-
-/**
- * A utility function that gets the computed size of a string when rendered in the browser.
- * @internal
- * @param text
- */
-function getTextSize(text: string): number {
-  let selection = d3.select("body").append("svg");
-
-  let width = selection
-    .append("text")
-    .attr("class", "tmp-text")
-    .text(text)
-    .node()!
-    .getComputedTextLength();
-
-  selection.remove();
-  return width;
-}
-
-/**
- * A utility function that maps the multiple levels of text detail to Annotations for later use.
- * @internal
- * @param config
- */
-function addToTextMaps<A extends Annotation, C extends Chart<any>>(config: {
-  annotations: A[];
-  chart: C;
-  textFn: (a: A, c: C) => string[];
-}): void {
-  for (const a of config.annotations) {
-    let text = config.textFn(a, config.chart);
-    textMap.set(a.id, text);
-    let thresholds = text.map((t) => {
-      let textSize = getTextSize(t);
-      return ((a.end - a.start) * config.chart.viewportWidth) / textSize;
-    });
-    thresholdMap.set(a.id, thresholds);
-  }
-}
-
-/**
- * An interface that defines the parameters for a call to the text rendering function.
  */
 export interface TextConfig<A extends Annotation, C extends Chart<any>>
   extends GlyphConfig<A, C> {
@@ -120,20 +40,11 @@ export interface TextConfig<A extends Annotation, C extends Chart<any>>
    * https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/alignment-baseline
    */
   alignmentBaseline?: GlyphProperty<A, C, string>;
-  /**
-   * A callback to extract a list of text to display from the represented Annotation object. It is a list of text
-   * because TextGlyphs can display varying length text depending on how much room is available at the Chart's
-   * current zoom level.
-   * @param a
-   * @param c
-   */
-  textFn: (a: A, c: C) => string[];
   initializeFn?: (this: TextModifier<A, C>) => void;
   zoomFn?: (this: TextModifier<A, C>) => void;
 }
 
 /**
- * An interface that defines the parameters to instantiate a TextModifier.
  * @internal
  */
 export type TextModifierConfig<
@@ -142,7 +53,6 @@ export type TextModifierConfig<
 > = GlyphModifierConfig<A, C> & TextConfig<A, C>;
 
 /**
- * A class that manages the styling and positioning of a group of text glyphs.
  * @internal
  */
 export class TextModifier<
@@ -158,7 +68,6 @@ export class TextModifier<
 
   constructor(config: TextModifierConfig<A, C>) {
     super(config);
-    addToTextMaps(config);
     this.fillColor = config.fillColor || "black";
     this.strokeColor = config.strokeColor || "none";
     this.textAnchor = config.textAnchor || "start";
@@ -177,17 +86,11 @@ export class TextModifier<
     this.applyFontFamily();
     this.applyFontStyle();
     this.applyAlignmentBaseline();
-    this.applyText();
   }
 
   defaultZoom(): void {
     this.applyX();
     this.applyY();
-    this.applyText();
-  }
-
-  applyText(): void {
-    this.selection.text((d) => selectText(d.a, d.c));
   }
 
   applyTextAnchor(): void {
@@ -213,31 +116,4 @@ export class TextModifier<
   applyAlignmentBaseline(): void {
     this.applyStyle("alignment-baseline", this.alignmentBaseline);
   }
-}
-
-/**
- * This renders a list of Annotation objects as text in a Chart.
- * @param config
- */
-export function text<A extends Annotation, C extends Chart<any>>(
-  config: TextConfig<A, C>
-): d3.Selection<SVGGElement, string, any, any> {
-  let selector = config.selector || generateId("soda-text-glyph");
-  let internalSelector = selector + "-internal";
-
-  let binding = bind<A, C, SVGTextElement>({
-    ...config,
-    selector,
-    internalSelector,
-    elementType: "text",
-  });
-
-  let modifier = new TextModifier({
-    ...config,
-    selector: internalSelector,
-    selection: binding.merge,
-  });
-  config.chart.addGlyphModifier(modifier);
-
-  return binding.g;
 }
